--- src/main/java/com/example/movierental/model/Admin.java
@@ -1,16 +1,0 @@
-package com.example.movierental.model;
-
-//Admin Class
-//
-//@author Jack Murphy - 18254268
-
-public class Admin extends User{
-
-    public Admin() {
-        //empty constructor
-    }
-
-    public Admin(int userID, String username, String password, boolean banned) {
-        super(userID, username, password, banned);
-    }
-}
